handle caching at the provider level
this commit adds a class for cached json based rest apis to replace the
(quite ugly) global url based caching and to remove some code
duplication across imgur and gfycat.

It also improves error handling as http return codes are now checked and
raised as an exception if they don't indicate a success.