Update post metadata layout
Links to commend and author are now wrapped into buttons located
on right side below the post. Also the author and post categorie
are located on the lower left side of post container. This gives
the post a more minimalistic view and let the user focus the post
and title first.

Main changes:
* Icon for comment link
* Icon for source link
* author below the post
* reddit categorie below the post
* padding of 5px to seperate post and metadata