update Frontend to JSON generated content
Since the server is sending JSON there is now a way to use this
information to build up the site. The site content itself is now
completly based on the JSON content. "redditgag.js" is spereated
in two main functionalities:
1. Tracking if the user scrolled to the botton
2. If so, then generating a new JSON request and building up the
   new content.

Since we can append content to the site it is not longer needed to
generate the side content static inside the backend. That's the
reason why I removed all the content from "index.html" and shited
it inside "redditgag.js". I tried to translate one to one the
template function into javaScript code. Please verify that this
code is corret.
In addition there will be a loading animation created whilte wait-
ing for the incoming data.

I was not realy sure weather I can play arround with the limit=25
value from the reddit API. I think it would be a good idea to set
this limit to a lower value to prevent long loading times of big
images since they may interrupt the scrolling flow of the user.