move exception sorries to providers.get_embeddable
This allows providers to focus on returning their actual content instead
of a SORRY post.

However, returning SORRY because of an unsupported link type that the
provider should be responsible for is also pretty common. This could
probably be realized as a NotSupportedException. Alternatively, a
solution that moves the domain check out of the individual providers
might also be worth considering (since it would also happen pretty much
everywhere)

This commit also adds TODOs for deploying the app and adding a few unit
tests.