add sorry post if ajax request fails
To understand the code, remeber that `@` is a shortcut for `this.` in
Coffeescript.

The `.success()` callback (the function passed directly to `getJSON()`
is deprecated and all it did can be done in the `.done()` callback.

The response data (or the error message) is stored in the `response`
property of the jqxhr's context (the thing that `this` refers to in
javascript).