Add Tests for providers that query external apis
Good luck to whoever is trying to understand how this works in a few
months.

The idea behind the tests is that I didn't want to write a lot of actual
code to test a bunch of providers that basically all work the same way:

 * some url goes in
 * the provider extracts some identifying marks from it,
 * queries some external api with them
 * and maps the result of the api call to some sensible representation

This is why there now is a data driven test that just mocks out the
external api call and checks that the arguments to it and the resulting
embeddable match their expected values.

The test data is stored in app/test_data/app.providers.*.json as a list
of json objects containing the keys:
- url -> the url that should be embedded by the provider
- api_params -> the expected set of features of the url that the
                provider will use to query the api
- api_result -> the expected result of the api call
- expected -> the expected final result of the call to provider.embed

I also added a management command to populate the test vectors based on
the current results for a given url. It replaces the entire file though
so be careful not to run it if you have unstaged changes to the test
data or have just added a new test case that is expected to fail.